[css2] Issue 186 (QA guidelines by Tim Boland)
Ad req 1:

Changed 3.2 "Conformance" to "UA Conformance" to make it clear what is
required to conform.

Ad req 3:

Added that an authoring tool is a kind of user agent.

Ad req 5:

Added a definition for "property."

Changed "support one or more of the media types" to "recognize one or
more of the media types" and "supported media" to "recognized media".

Not changed anything to clarify what "attempt to retrieve" means. I
think it is reasonably clear. If a UA under test doesn't apply a style
sheet it is supposed to apply and yet another program on the same
machine at approximately the same time succeeds in reading the style
sheet, then the UA is probably not conformant. The more often this
happens, the higher the probability that the UA is not conformant.

Ad req 6:

Nothing changed. The conformance section already defines a few classes
of UAs and the various CSS profiles (TV, print, mobile) define more.

Ad req 7, 8, 9, 10:

Nothing changed.

Ad req 11:

Nothing changed. Is it necessary to say what CSS is not (viz., not
extensible)?

It might be possible to add a paragraph somewhere that says that,
although the CSS syntax is designed to accomodate future extensions,
users and authors cannot extend CSS, because it has no mechanisms to
define the semantics of new features (other than the W3C
Recommendation process).

Ad req 12:

Changed 18.2 "CSS2 System Colors" to "System Colors" (in ui.src)

Ad req 13:

Nothing changed. (The only deprecated feature in CSS 2.1 is the
'aural' media type and the appendix clearly says that it doesn't
affect conformance.)

--HG--
extra : convert_revision : svn%3A73dc7c4b-06e6-40f3-b4f7-9ed1dbc14bfc/trunk%402510