[css2] Updated references as requested in http://www.w3.org/2002/09/wbs/33280/css21pr/results
Not changed UAAG 1.0 to 2.0, because latter is not a REC yet.
Updated UAX9 to point to Unicode 6.0.0.
Updated Unicode itself to point to 6.0.0.
(XML points to Unicode 5.0.0 but that seems fine.)
Updated ref for CSS Color to the PR version.
Updated ref for Selectors to the PR version.
Updated ref for CSS Speech to April WD.
Updated MathML to 3.0.
Updated Media Queries to CR version.
Not changed SVG 1.1 to 2nd edition, because it is not REC yet.
Updated ref for  WCAG 1.0 to 2.0.
Updated ref for XML:ID to REC version.
Updated ref for XML Namespaces to 3rd edition.

--HG--
extra : convert_revision : svn%3A73dc7c4b-06e6-40f3-b4f7-9ed1dbc14bfc/trunk%403212