Explain the definition of backface-visibility
This is editorial, so I'm doing it without discussion as we agreed.  It took me
a little while to puzzle out why the definition made sense, because I was
thinking of elements as having zero thickness, so I thought I'd share the logic
by putting it in the spec.  It's very mathematical, but since the definition is
completely matrix-based, that seems unavoidable.