[selectors4] Editorial: s/of its type/with the same expanded element name/
Currently, :first-of-type is described as …

> an element that is the first sibling of its type in the list of
> children of its parent element.

And similarly for :last-of-type.

"of its type" is somewhat ambiguous: it does not say if the namespace
URIs have to match or not.

:nth-of-type() and :nth-last-of-type() however both use a more precise
wording: "with the same expanded element name". css3-namespace defines
the expanded name as the tuple of the namespace URI and local name.

This changes :first-of-type and :last-of-type to the more precise wording,
and links each occurrences of "expanded name" to its definition
in css3-namespace. (Which Selectors already normatively references.)

--HG--
extra : rebase_source : 9e74406b4e24df438b23e263db3272043f2d1fb4