[css3-page] Rewrite 6.3.2 Margin Box Variable Dimension (again)
While in the previous iteration the layout of page-margin boxes was
similar to floats, it is now closer to normal blocks.

The biggest change is that boxes with 'width: auto' tend to occupy all
of the available space, even if max-content is smaller than that.

Also, they do not always get at least min-content anymore. If a sentence
has one very long word that has to overflow, there is no reason that
other lines should be allowed to overflow too.

Lastly, this new algorithm should be saner when mixing auto and non-auto
widths.