[css-*] Replace s/rule ?set/style rule/, except in grammars.
Having three names for the same concept is silly.
"Rule set" is confusing as it implies multiple rules,
while it in fact refers to a single rule.
CSS OM and other specs already use "style rule".

This changes "rule set" and "ruleset" to "style rule",
except when used in a grammar.
(The normative definiton of the Core Grammar is in 2.1
which is harder to change.)

This is missing a few changes in css-fonts,
since its source is not version-controlled in this repository.