[css-syntax] Not all qualifed rules contain declarations.
Similarly to at-rules, the block of qualified rules is now just a block,
and is not parsed as a declarations list at the Syntax level.

CSS already has qualified rules that are not style rules
inside @keyframes, even though they still contain declarations.

With this change, a future CSS spec can use the qualified rule syntax
without the declaration syntax, without having to change this spec.

Parsers for existing qualified rules conceptually just need to call
"Parse a list of declarations" on the rule’s block.