Clarification for an example
However, the span element is not in fact broken in two for the purposes of styling.

If the span had a border or background image, and the browser acted as if the element were closed and reopened, a no-repeat background image would be displayed twice. or, if a border were present, the border would be closed and reopened.