[css-logical-props] Explain logical box model properties in a little more detail, and give them a proper toplevel section heading.
This adds an explanation of how logical and physical properties cascade
together.

It also adds an additional open issue about whether getComputedStyle()
should accept logical properties.  (In Gecko's current implementation,
I believe it does not do so, since it seemed easier to add later if
we want to add it than to remove later if we don't.)

This also makes the logical properties have their own toplevel section
rather than being a subsection of the logical page classifications,
which seems like it was a mistake.

I'm just committing this per
https://log.csswg.org/irc.w3.org/css/2015-07-21/#e573158