[css-ui-4] Comment out examples that use prefixes
webkit prefixed animations were used in live examples to make sure they
were animated in Safari as well. Since there is no agreement in the
publication process that this is an OK thing to do, I am temporarily
commenting out the examples until we either agree that this is OK, or
agree on an alternative way to write these examples.