[cssom-view] Change how {Element,Range}.getBoundingClientRect() handle empty rects
The special handling of the first rect was to return something
useful when all rects are empty, per feedback from roc:
https://lists.w3.org/Archives/Public/www-style/2009Jul/0149.html
https://github.com/w3c/csswg-drafts/commit/ccff98328b1c0019e953f82d98b52e32f2b82681

The spec did not match Gecko, and when Blink implemented the spec[1]
they received feedback[2] that it was confusing when the first rect
is empty but a later rect is not empty.

Change the spec to only return the first rect if all of them are
empty, and otherwise ignore the empty rects. This matches Gecko
and soon Blink. [3]

Fixes http://www.w3.org/mid/CAN9ydbUOzz4BJieAa-4ofcxFv25xtZ+qbairxwVh7wVsF=VS1w@mail.gmail.com

[1] http://crbug.com/574363
[2] http://crbug.com/606662
[3] http://www.w3.org/mid/CAN9ydbXN290=_Mg4Ak1mn7yvsvDLOSzjvZ-1=ST5iMJA=ykPvA@mail.gmail.com