[css-flexbox] Replace 'Animatable' descriptions with 'Animation type'
With regards to 'flex-basis', I wasn't sure if "Animation type: length,
percentage, or calc; otherwise discrete" was more clear about the case
when the value is 'content'. However, we should probably just make sure
the general wording for resolving animation types covers this case, i.e.
"if you can't stuff the value in the animation type, it's discrete".