[css-transitions-2] Simplify and correct generalized transition event dispatch
I decided to just ignore the initial playback rate of the animation
since it was getting far too complicated. To do it properly, you'd
actually have to change the initial set of events that get dispatched
based on the playback rate at a fairly arbitrary moment. Instead, the
revised behavior just assumes a positive playback rate for the
elapsedTime for the initial set of events.