Add additional any-* consideration note to pointer/hover
The rationale given in any-pointer/any-hover about not relying
exclusively on any-* is equally valid in reverse here. This does not
imply that authors must design for "lowest common denominator", but
simply opens up the possibility that they may wish to keep non-primary
inputs in mind when deciding on their layout/functionality.