[css-conditional] use standard whitespace convention for grammar terms again
This reverts the remainder of
https://hg.csswg.org/drafts/rev/34b185ae3bac
which was partially reverted in
https://hg.csswg.org/drafts/rev/3091bbbdc67a

This revert restores this grammar to follow the standard convention that
all terms consume the whitespace at their end.  When there was mandatory
whitespace around 'and', 'or', and 'not' in supports conditions, we
couldn't follow that convention because the mandatory whitespace might
be consumed as part of a term in which the whitespace was optional.  But
now that there is no more mandatory whitespace, we should follow the
usual convention.

I made the edits manually, and then checked that it matched the state of
the grammar before the initial change (which it does).