[css-timing] Drop before flag from frames() timing function
The before flag was added primarily for step-start because at input progress = 0
the output should be the top of the step but during the delay phase it should be
the bottom of the step.

As an extension of that, we said that if you used the iterationStart property to
offset the beginning of the animation such that it lines up exactly with one of
those steps, then we should behave in the same way: i.e. bottom of the step
during the delay, top of the step otherwise.

 Original discussion:
 https://lists.w3.org/Archives/Public/public-fx/2013OctDec/0233.html

 Subsequent discussion:
 https://old.etherpad-mozilla.org/yhmUQAEfpo

 Issue:
 https://github.com/w3c/web-animations/pull/16

Originally I applied the same behavior to frame timing functions. However, they
don't have the same issue steps occurring at progress=0. Furthermore, for the
iterationStart case, suppose you have frames(4) and set iterationStart to 0.25,
then I think during the delay phase you actually do expect that animation to
show the second frame. That is, you don't want the step down behavior. Instead
you're effectively saying, "I've asked for 4 frames but I want to start from the
second frame (presumably even when I'm filing."

So I've dropped the before flag behavior for the frames() timing function.