Remove the spatnav-container property
While it may be nice to have and doesn't cost much on top
of the processing model, it is non essential, and can be deferred to
level 2.

Keeping it in the spec for now was distracting:
not only were there open issues (especially naming) related to it,
but distracted readers from the essential point of this specification.