[css-animations] Change AnimationEvent elapsedTime member from 'float' to 'double'
As per the web-idl spec, float should only be used if there is a
specific reason, as double much more closely matches the ECMAScript
Number type. In the case of AnimationEvent there does not appear to be
such a reason.

Furthermore, switching to double allows the elapsedTime to reach higher
values without overflowing. A rough calculation (with some assumptions
and possible errors) suggests that this change will raise the celing
from ~4.5 hours to years.