[css-transforms-2] Explicitly reset the temp matrix in matrix recomposition.
The idea of the blocks (IIUC) are to reset `temp` to be an identity matrix in
case it's been used before.

However if only skew[0] and skew[2] are non-zero, this code didn't manage to
accomplish it.

Make it so that the reset is explicit after setting a matrix member instead of
try to undo all other possible mutations that have happened before.