Add 2nd single-quote around 'align-self/stretch'
There's a missing single quote here, which means this one instance of  align-self/stretch gets quoted directly in the spec, with broken linkification around it. Here's the affected section if you're curious:
 https://drafts.csswg.org/css-align/#baseline-content-alignment

And here's a screenshot (good for 14 days):
  https://screenshots.firefox.com/V5tvNjkCu5juNkoF/drafts.csswg.org

Looks like we're just missing a single quote here. This pull request adds the missing single-quote to fix it.