[css-text-3] Start notes with "Note:"
Prior to this change, some notes started with "Note:",
some started with "Note that",
and some started without any mention of it being a note.

This just makes everything consistent,
so that all notes can be readily identified.

I chose the "Note:" for over the "Note that" form as it
makes it less likely that we'll break something
in a later update to the spec to use less HTML and more bikeshed-flavored markdown,
since bikeshed turns "Note:" into notes even in the absence of
class=note.

Closes #3211