[cs-text-3] Close stale issue
This issue was open while we were still trying to solve the
'word-break:break-word' problem, with the assumption that
'overflow-wrap: break-word' could be made to change intrinsic sizing
computations, but that 'overflow-wrap: break-word; white-space: nowrap'
(or 'white-space: pre') would result into intrinsic computations not
being affected (because of white-space), but could still be made to
enable overflow-wrapping.

Since we've solved the problem differently, this suggestion is no longer
relevant and should be removed.