[css-ui] Add 'menulist-textfield' appearance value
This is an alias to 'none' for better web compat (compared to not
supporting the value at all or making it equivalent to 'menulist').

See https://bugzilla.mozilla.org/show_bug.cgi?id=1500423#c62

[css-ui] Add 'caret' appearance value

This value is supported cross-browser, but does different things.
In EdgeHTML/WebKit/Chromium, it is like 'none' but the border is
not painted. In Gecko, it is like 'none'.

Chrome Use counter for this value is ~0.266%.
https://www.chromestatus.com/metrics/feature/timeline/popularity/1355

Add menulist-text as an alias to none

Use <compat-none> and <compat-auto>