[css-flexbox-1] Use bracketed range notation
Added bracketed range notation for non-negative values ( [0,∞] )
in these properties:

- flex-grow
- flex-shrink

Also use it in the prose descriptions for `flex`,
and to replace the made up `<postive-number>` variable
in the examples of common patterns.

Created a new "Changes since the 19 November 2018 CR" section
to include a mention of the update.

The syntax for `flex-basis` will be updated by reference
when `width` is updated.