[css-text] Clarifications on "hanging"
Separate the notion of hanging in general, from it's specific
application to hanging-punctuation, and provide reusable anchor
terminology that distinguishes unconditional hanging from conditional hanging.

Also, be explicit about the effects on intrinsic sizes.

Closes #3784

Note: there are open issues and pending edits on how preserved
white-space at the end of a line is supposed to hang. This edit does not
solve these, and keeps the current behavior as it is. By making the
definitions of hanging reusable, it is a useful step towards solving
these as well.