Add note re scripts to text-decoration-skip-ink
Given that there is not a single canonical algorithm for determining the boundaries of script runs in arbitrary text, and given that some implementor want to retain the freedom to make adjustments based on user feedback, I don't think we can or should currently specify precise, normative requirements for how `text-decoration-skip-ink` behaves down to the level of every individual character.

However, the spec could usefully include a (non-normative) note to offer guidance to implementors. Here is some proposed text for consideration by the editors.