[css-text-3] replaced elements and atomic inlines with non default line-break
This is a follow up to #4576

This makes replaced elements and atomic inlines with non default
line-break behave the same as characters with line-breaking class ID.
This gives the expected behavior for images used as “gaiji” and as
emoji.

This cannot be the default behavior for compat reasons,
but having it as an opt-in based on the rarely used line-break property
should be safe.