[CSS2] Make TOC use first id in <h_>
This drops the requirement for it to be on a span (so e.g. if it is
on a dfn it will now be picked up). This also makes us strip *all*
tags within headers when generating the TOC to avoid copying the id
and anything else (like a dfn!).