[CSS2] eliminate new line characters within HTML tags
This is done for the same reason as the previous commit (both Bikeshed
conversion tooling does this, so this limits the diff in that single
commit making blame more workable, and because Bikeshed does not handle
tags split across multiple lines). Note that this doesn't deal with new
lines within attributes yet.