[CSS2] put source files through a html5lib parse/serialize cycle
This is, similar to parent commits, prior to the Bikeshed conversion
to make more easily ignored from blame (plus also is easier to
review).

Most of this just gets rid of ambiguous ampersands and makes quoting
of elements consistent (the choice of quotes is done to minimize the
size of this diff).