[CSS2] get rid of q0 anchors
The q0 anchors date back to the original 2011 publication of CSS 2.1 and
are all applied to the top-level headings within each page. As such,
they're essentially unused (because everywhere just links to the top of
the page). Additionally, the q0 anchors were removed when CSS 2.1 was
edited in place in 2016 (see
https://github.com/w3c/csswg-drafts/issues/2551).