Fix intermediate color in color-mix(red yellow 65%) example.
The main bug was reported both in a comment in the spec and in
https://github.com/w3ctag/design-reviews/issues/526#issuecomment-649100113

This also fixes dropping of a "9" digit in the red's chroma midway
through the calculations.

Note that the correct intermediate is well out of sRGB.  I just chose to
use it anyway and add a note about such cases.