[css-color-5] Adjustments to edge cases and set notation for hue interpolation types.
This makes a few adjustments to the definitions of hue interpolation
types to follow up on the discussions in #4735.

First, it adjusts the pseudo-code so that for 'shorter' and 'longer',
angles that differ by 180 degrees will always go in the decreasing
direction, as described in
https://github.com/w3c/csswg-drafts/issues/4735#issuecomment-650254063

Second, it adjusts the set notation to match the pseudo-code.  (They
were not matching even prior to this change.)  The set notation
previously seemed to assume that there were absolute-value functions
present that were not actually there.  However, given the asymmetry of
always going in the decreasing direction, it's more correct to write it
without absolute values.