[css-fonts] Add font-size: math, math-level and math-style
This imports the proposal from MathML Core to manage font scaling in math
formulas using a new 'math-level' property. This mechanism involves a new
'math-style' property that describes compactness of formulas as well as
a new 'math' keyword to specify relative font-size change using the a
scale factor calculated from the 'math-level' changes.

https://github.com/w3c/csswg-drafts/issues/5389
https://github.com/w3c/csswg-drafts/issues/5387