Update Overview.bs
Changed initial from auto to none for text-decoration-skip-ink-property. Reason: the auto setting leaves the default behaviour at the discretion of implementer which is never a good thing - a standard should be well defined and unambiguous, and may change the look of sites without knowledge of their creators and owners which is not a good idea either.