[css-contain-2] Allow exceptions to size containment
Normally, properties applied on a size-contained element itself do
affect the size of the element. There may be an interest in exempting
some properties, like grid track sizing properties, from that, but it is
currently undecided. For now, just spec the simple behavior without
exemptions, but allow exemptions in other specs to exist.