[css-grid-1] Specify height in example 8 for #grid (#5968)
I would specify the height property in example 8 of the Grid Areas
chapter for `#grid`. This makes it easy to see the differences between
`align-self: start;` and `align-self: end;` for newcomers. Otherwise,
it's may not so clear how `align-self: start;` and `align-self: end;`
styles work if you try to run this snippet.

[3.3. Grid Areas](https://www.w3.org/TR/css-grid-1/#grid-area-concept)