[css-fonts-5] Broaden language about @font-face metrics descriptors' normal value
The original text was overly specific. In particular, `normal` values don't always
need to come directly from the font file. `normal` is intended to just do the same
thing as what happens today, without any of these metrics descriptors.

Both editors agree that this is just an editorial fix, and can just be committed
without an issue or WG discussion.