Fix incorrect CSS in text-decoration example
The paragraph preceding the example for the text-decoration property in both CSS Text Decoration Module Level 3 and Level 4 says:

> The following example underlines unvisited links with a solid blue underline in CSS1 and CSS2 UAs and a navy dotted underline in CSS3 UAs.

However, the example uses a type selector of `link`. The `link` element is not used for hyperlinks, but instead to link their document to other resources, more typically used in the `<head>` of the document. The `a` element is used for hyperlinks. As a result, the example CSS does not do what is described.

Looking at the history, it looks like the selector was originally `:link` but the colon was mistakenly removed in b32b070, a commit which was intended to only change whitespace. The incorrect example was then copied into the Level 4 spec.

Restore the original `:link` selector, which would correctly targets unvisited links.