[cssom-1] Refactor CSSStyleSheet constructor and replaceSync to allow steps to be referenced from HTML spec (#6411)
The CSS module scripts feature is being specified in HTML using several algorithms defined in cssom-1 on the CSSStyleSheet interface. In the review of the CSS module scripts HTML spec PR whatwg/html#4898, there was a concern that referencing the CSSStyleSheet methods directly implies that if these methods were overridden by user script, the UA should call the user script instead.

This change pulls the steps for the `CSSStyleSheet` constructor and `replaceSync` method into separate algorithms so that they can be safely referenced from the HTML spec.