[css-scrollbars] Simplify abstract
Broaden the scope to cover everything in the module.

Also avoids drawing attention to IE5.5: it did lead the way on this
topic, but that mention has made various people see this specification
in a negative light, assuming that what it's doing must be antiquated if
it's coming from something this old.