[css-font-loading] Add IDL for discovery of information about a font
This is for https://github.com/w3c/csswg-drafts/issues/520.

The [resolution](https://github.com/w3c/csswg-drafts/issues/520#issuecomment-387799523) is:

> RESOLVED: [add this functionality to] css-font-loading, level TBD

I added this to level 3, because level 4 doesn't exist. If the editor prefers this go in a level 4, that's no problem.

The GitHub thread hasn't come to a conclusion about what should go in the `FontFaceFeatures` interface, so I left it blank for now.