[web-animation-2] Making currentTime and startTime CSSNumerish (#6479)
The animation attributes currentTime and startTime were nullable doubles and represented times in units of milliseconds. These units do not align well with progress based animations where it is more natural to work in percentages.

With this change, progress based animations expect and report times as percentages. When setting current or start time, the value may be expressed as a double with implicit units of milliseconds, or as a CSSNumericValue with duration units ('s' or 'ms') when the animation is associated with a time based timeline.

Co-authored-by: Kevin Ellis <kevers@chromium.org>