[css-transitions-1] Define before/after-change style in terms of base
 - Introduce "base style", which is the computed style as it would be
   without animations/transitions.
 - Define before/after-change style in terms of that base style
   (at two different points in time).
 - There should be no need for the "completed transition" set anymore,
   since the before/after-change styles on a child are blind to any
   effects from transitions/animations happening on the parent.
 - Maintain the behavior that transitions can not start for properties
   affected by animations. Removing this behavior was considered,
   but was deemed impossible from use-counter data showing that ~5%
   of sites may rely on this behavior.

This change stems from Container Queries, which blurred the line
between the style and layout steps of the frame. The spec currently
defines the before-change style as the style of the previous
style change event, but with any animation effects updated to
current time. This is difficult, because it persisting the before-change
world in some form that is responsive to what the (newly updated)
animations actually do. (For example, we'd need to re-resolve var(),
font-relative units, be able to handle revert/revert-layer, etc.)
With Container Queries, this problem becomes substantially worse,
because now we'd also need to maintain the "before-change layout tree",
which in practice is gone at the time we need it.

#6398