[web-animations-2]  Update ComputedEffectTiming (#6712)
This PR fixes descriptions in ComputedEffectTiming to remove ambiguity over the return value types. Specifically, effects linked to animations with timelines that have an unresolved duration should return times expressed as doubles in milliseconds and not as CSSNumericValues. Only progress based animations should return their time values expressed as CSS.percent.

The progress attribute was previously changed to be of type CSSNumberish but without a description change. This property is a fraction and not a time. Thus, it should not have been changed. The change is reverted in this PR.