[css-transforms-2] Define that backgrounds/borders/decorations of element establishing 3D Rendering Context are rendered at z=0 in its 3D scene.
This defines that backgrounds, borders, and other box decorations of an
element establishing a 3D Rendering Context are rendered at z=0 in its
3D scene, rather than behind its 3D scene.

This matches the current behavior of Chromium and Gecko.  While WebKit
implements the spec as written, it also doesn't match the current spec's
definition of 3D Rendering Contexts that covers up the difference, so
implementing the spec as written today may not be web-compatible.

Tests in https://github.com/web-platform-tests/wpt/pull/31375 ,
particularly -004 (although -005 is more relevant to WebKit).

Fixes #6238.