[css-overflow-3] Clarify compatibilty value and property
The -webkit- prefixed compat part of line-clamp and related properties
aren't meant to be anything out of ordinary in terms of property
mechanics, cascading, OM, etc. They're defined in a separate section do
deemphasize them, but otherwise behave normally. The previous
description was confusing about that, so this clarifies that.

See #6842