[css-ui-4] Support cursors using a limited version of image-set()
The more generic <<image>> production continues to be allowed but not
required.

For clarity, since image-set() is a list of images, and since it would
itself be used in a list of cursors, this commit rearranges the text
editorially a little, in order to limit the confusion about which list
we're talking about.