Audit specifications for correct handling of overflow:clip.
This set of changes is the result of auditing all specification
Overview.bs other than CSS2 and css-overflow-* for correct handling of
overflow:clip, since many specifications have historically referenced
concepts like "overflow values other than visible".

It is related to #6212, but doesn't actually fix it, because it's really
only a bug in CSS2.